% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emidata-package.R
\docType{package}
\name{emidata-package}
\alias{emidata}
\alias{emidata-package}
\title{emidata: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
